Refactor performance code
